+++ app/company/location/page.tsx
@@ -1,4 +1,4 @@
-import CompanyDetailTable from "@/app/components/CompanyAddressTable";
+import CompanyDetailTable from "@/app/components/company/CompanyAddressTable";
 
 export default function CompanyLocation() {
   return <CompanyDetailTable />;
