+++ app/components/company/CompanyDetailTable.tsx
