+++ app/components/company/historyCard.tsx
